fix(build): separate GCS connection setup to avoid chunk collision
