+++ apps/sim/lib/uploads/providers/gcs/client.ts
@@ -1,14 +1,13 @@
 import type { Readable } from 'node:stream'
-import type { Storage } from '@google-cloud/storage'
 import { createLogger } from '@sim/logger'
 import { generateId } from '@sim/utils/id'
-import { env } from '@/lib/core/config/env'
 import {
   assertKnownSizeWithinLimit,
   readNodeStreamToBufferWithLimit,
 } from '@/lib/core/utils/stream-limits'
 import { GCS_CONFIG } from '@/lib/uploads/config'
 import { isObjectNotFoundError } from '@/lib/uploads/core/errors'
+import { getGcsClient } from '@/lib/uploads/providers/gcs/connection'
 import type {
   GcsConfig,
   GcsMultipartPart,
@@ -32,67 +31,6 @@ const logger = createLogger('GcsClient')
 const GCS_XML_API_HOST = 'https://storage.googleapis.com'
 const GCS_MULTIPART_STAGING_PREFIX = '.sim-multipart/'
 const GCS_MULTIPART_UPLOAD_ID_METADATA_KEY = 'sim-upload-id'
-
-let _gcsClient: Storage | null = null
-
-/**
- * Reset the cached GCS client. Only intended for use in tests.
- */
-export function resetGcsClientForTesting(): void {
-  _gcsClient = null
-}
-
-interface GcsInlineCredentials {
-  client_email: string
-  private_key: string
-  project_id?: string
-}
-
-/**
- * Parse the inline service-account JSON from `GCS_CREDENTIALS_JSON`.
- * Returns null when the variable is unset (Application Default Credentials).
- * @throws Error when the variable is set but not valid service-account JSON
- */
-export function parseGcsCredentials(): GcsInlineCredentials | null {
-  if (!env.GCS_CREDENTIALS_JSON) return null
-
-  let parsed: unknown
-  try {
-    parsed = JSON.parse(env.GCS_CREDENTIALS_JSON)
-  } catch {
-    throw new Error('GCS_CREDENTIALS_JSON is not valid JSON')
-  }
-
-  const credentials = parsed as Partial<GcsInlineCredentials>
-  if (!credentials.client_email || !credentials.private_key) {
-    throw new Error('GCS_CREDENTIALS_JSON must contain client_email and private_key')
-  }
-
-  return credentials as GcsInlineCredentials
-}
-
-export async function getGcsClient(): Promise<Storage> {
-  if (_gcsClient) return _gcsClient
-
-  const { Storage } = await import('@google-cloud/storage')
-  const credentials = parseGcsCredentials()
-
-  _gcsClient = new Storage({
-    ...(env.GCS_PROJECT_ID || credentials?.project_id
-      ? { projectId: env.GCS_PROJECT_ID || credentials?.project_id }
-      : {}),
-    ...(credentials
-      ? {
-          credentials: {
-            client_email: credentials.client_email,
-            private_key: credentials.private_key,
-          },
-        }
-      : {}),
-  })
-
-  return _gcsClient
-}
 
 /**
  * Get an OAuth2 bearer token for authenticated XML API requests
